Module OLD.Group

This module implements a hierarchical version of cobol data items.

type t' =

| Renames of {

name : Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc;

targets : t list;

}

| ConditionName of {

name : Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc;

values : Cobol_ptree.condition_name_value list;

target : t;

}

| Constant of {

name : Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc;

value : Cobol_ptree.constant_value Cobol_common.Srcloc.TYPES.with_loc;

constant_item_descr : Cobol_ptree.constant_item_descr;

}

| Elementary of {

name : Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc option;

data_item : Cobol_ptree.data_item_descr;

}

| Group of {

name : Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc option;

elements : t list;

data_item : Cobol_ptree.data_item_descr;

}

The type of hierarchical data items from cobol, used to describe different sections of the data division such as working storage section or linkage section.

and t = t' Cobol_common.Srcloc.TYPES.with_loc

val pp_t' : Ppx_deriving_runtime.Format.formatter -> t' -> Ppx_deriving_runtime.unit

val show_t' : t' -> Ppx_deriving_runtime.string

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val compare_t' : t' -> t' -> Ppx_deriving_runtime.int

val compare : t -> t -> Ppx_deriving_runtime.int

val pp_data_group_list : Stdlib.Format.formatter -> t list -> unit

val of_working_item_descrs : Cobol_ptree.working_item_descr Cobol_common.Srcloc.TYPES.with_loc list -> t list Cobol_common.Diagnostics.with_diags

Convert a list of located tworking_item_descr_entry to a list of t t