Module OLD.Qualmap

type 'a t

val pp : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a t -> Ppx_deriving_runtime.unit

val show : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a t -> Ppx_deriving_runtime.string

val empty : 'a t

val find : Cobol_ptree.qualname -> 'a t -> 'a

find qualname map returns the unique element e of map that can be qualified with qualname. Raises Not_found if the element is not unique, or does not exists in map.

val find_opt : Cobol_ptree.qualname -> 'a t -> 'a option

find_opt qualname map returns Some elt if elt is the unique element of map that can be qualified by qualname. It returns None if there is no element or the uniqueness is not respected.

val find_all : Cobol_ptree.qualname -> 'a t -> (Cobol_ptree.qualname * 'a) list

find_all qualname map returns all the elements that can be qualified with qualname in map with their fully qualified name.

val find_full_qualname : Cobol_ptree.qualname -> 'a t -> Cobol_ptree.qualname

find_full_qualname qualname map returns the full name of qualname if it is unique

val find_full_qualname_opt : Cobol_ptree.qualname -> 'a t -> Cobol_ptree.qualname option

find_full_qualname_opt qualname map returns Some full_name with full_name being the fully qualified name of qualname if it exists, None otherwise

val add : Cobol_ptree.qualname -> 'a -> 'a t -> 'a t

add qualname value map returns map with value bound to the key qualname. qualname is assumed to be the fully qualified name of value.

val iter : (Cobol_ptree.qualname -> 'a -> unit) -> 'a t -> unit

iter f map iters over all the values of map.

val fold : (Cobol_ptree.qualname -> 'a -> 'b -> 'b) -> 'a t -> 'b -> 'b

fold f map init folds over all the key values bindings of map