Module Cobol_data.Picture

module TYPES : sig ... end

type t = TYPES.picture

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val compare : t -> t -> Ppx_deriving_runtime.int

val is_edited : TYPES.category -> bool

is_edited c indicates whether the given category represents an edited item

val data_size : TYPES.category -> int

data size (in "characters" --- probably with implemententor specific semantics)

val size : TYPES.category -> int

display size, after editions; corresponds to "size" in standards

val of_string : TYPES.picture_config -> string -> (TYPES.picture, (TYPES.error * (int * int)) list * TYPES.picture) Stdlib.result

module Make (Config : Cobol_config.T) (Env : TYPES.ENV) : sig ... end

val pp_meaning_of_precedence_index : TYPES.picture_config -> Stdlib.Format.formatter -> int -> unit

val unit_test : ?config:TYPES.picture_config -> expect:string -> string -> bool

Verifies that the picture string is interpreted as `expect`, i.e. the result of `pp_picture`. If not, displays the difference on stderr and returns `false`