Module Cobol_data.Types

Representation of COBOL data items

module NEL = Cobol_common.Basics.NEL

type 'a nel = 'a NEL.t

Literals

type alphanum_literal = {

alphanum_ptree : Cobol_ptree.alphanum;

alphanum_value : alphanum_value;
(*
differs from alphanum_ptree.given_str iff alphanum_ptree.hexadecimal holds
*)

}

and alphanum_value = string

and boolean_literal = {

bool_ptree : Cobol_ptree.boolean;

bool_value : boolean_value;

}

and boolean_value = {

bool_width : int;
(*
may be 0
*)

bool_bits : Z.t;
(*
irrelevant if 0-width
*)

}

and integer_literal = {

int_ptree : Cobol_ptree.integer;

int_value : integer_value;

}

and integer_value = Z.t

and fixed_literal = {

fixed_ptree : Cobol_ptree.fixed;

fixed_value : fixed_value;

}

and fixed_value = Q.t

and floating_literal = {

float_ptree : Cobol_ptree.floating;

float_value : floating_value;

}

and floating_value = {

float_significand : fixed_value;

float_exponent : int;

}

and literal_value =

| Alphanum_value of alphanum_value

| Boolean_value of boolean_value

| Integer_value of integer_value

| Fixed_value of fixed_value

| Floating_value of floating_value

| Zero_value

| Space_value

| Quote_value

| Low_value

| High_value

| All_alphanum_value of alphanum_value

Values that may come from literals only.

val compare_alphanum_literal : alphanum_literal -> alphanum_literal -> Ppx_deriving_runtime.int

val compare_alphanum_value : alphanum_value -> alphanum_value -> Ppx_deriving_runtime.int

val compare_boolean_literal : boolean_literal -> boolean_literal -> Ppx_deriving_runtime.int

val compare_boolean_value : boolean_value -> boolean_value -> Ppx_deriving_runtime.int

val compare_integer_literal : integer_literal -> integer_literal -> Ppx_deriving_runtime.int

val compare_integer_value : integer_value -> integer_value -> Ppx_deriving_runtime.int

val compare_fixed_literal : fixed_literal -> fixed_literal -> Ppx_deriving_runtime.int

val compare_fixed_value : fixed_value -> fixed_value -> Ppx_deriving_runtime.int

val compare_floating_literal : floating_literal -> floating_literal -> Ppx_deriving_runtime.int

val compare_floating_value : floating_value -> floating_value -> Ppx_deriving_runtime.int

val compare_literal_value : literal_value -> literal_value -> Ppx_deriving_runtime.int

Storage

type picture_config

type picture

type usage =

| Binary of picture

| Binary_C_long of signedness

| Binary_char of signedness

| Binary_double of binary_range_properties

| Binary_long of binary_range_properties

| Binary_short of binary_range_properties

| Bit of picture

| Display of picture

| Float_binary of {

width : [ `W32 | `W64 | `W128 ];

endian : Cobol_ptree.endianness_mode;

}

| Float_decimal of {

width : [ `W16 | `W34 ];

endian : Cobol_ptree.endianness_mode;

encoding : Cobol_ptree.encoding_mode;

}

| Float_extended

| Float_long

| Float_short

| Procedure_pointer

| Function_pointer of Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc

| Index

| National of picture

| Object_reference of Cobol_ptree.object_reference_kind option

| Packed_decimal of {

picture : picture;

with_sign_nibble : bool;

}

| Pointer of Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc option

| Program_pointer of Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc option

and signedness = {

signed : bool;

}

and binary_range_properties = {

signed : bool;

digits : int option;

scaling : int;

}

type data_storage =

| File of Cobol_ptree.name Cobol_common.Srcloc.TYPES.with_loc

| Local_storage

| Working_storage

| Linkage

type length_variability =

| Fixed_length

| Variable_length

Records

type record = {

record_name : string;

record_storage : data_storage;

record_item : item_definition Cobol_common.Srcloc.TYPES.with_loc;

record_renamings : record_renamings;

}

and item_definitions = item_definition Cobol_common.Srcloc.TYPES.with_loc nel

and item_redefinitions = item_definition Cobol_common.Srcloc.TYPES.with_loc list

and item_definition =

| Field of field_definition
(*
for data items without OCCURS clause
*)

| Table of table_definition
(*
for data items with an OCCURS clause
*)

and field_definition = {

field_qualname : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option;

field_leading_ranges : table_range list;

field_offset : Cobol_data__.Data_memory.size;
(*
offset w.r.t record address
*)

field_size : Cobol_data__.Data_memory.size;

field_layout : field_layout;

field_length_variability : length_variability;

field_conditions : condition_names;
(*
Named conditions on the value of this field.
*)

field_redefines : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option;
(*
Set iff this field is a redefinition. In that case this field appears inside item_redefinitions of the item it redefines. Later, we may create instead a item_redefinition type.
*)

field_redefinitions : item_redefinitions;
(*
List of alternative definitions for this field
*)

field_has_definition_issues : bool;

}

and field_layout =

| Elementary_field of {

usage : usage;

init_value : literal_value Cobol_common.Srcloc.TYPES.with_loc option;

}

| Struct_field of {

subfields : item_definitions;

}

and table_definition = {

table_field : field_definition Cobol_common.Srcloc.TYPES.with_loc;

table_offset : Cobol_data__.Data_memory.size;

table_size : Cobol_data__.Data_memory.size;

table_range : table_range;

table_init_values : Cobol_ptree.literal Cobol_common.Srcloc.TYPES.with_loc list;

table_redefines : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option;

table_redefinitions : item_redefinitions;
(*
List of alternative definitions for the full table. Note that by default the typechecker generates a warning on table redefinition.
*)

table_has_definition_issues : bool;

}

and table_range = {

range_span : span;

range_indexes : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc list;

}

and span =

| Fixed_span of fixed_span

| Depending_span of depending_span

| Dynamic_span of dynamic_span

and fixed_span = {

occurs_times : int Cobol_common.Srcloc.TYPES.with_loc;

}

and depending_span = {

occurs_depending_min : int Cobol_common.Srcloc.TYPES.with_loc;

occurs_depending_max : int Cobol_common.Srcloc.TYPES.with_loc;

occurs_depending : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc;

}

and dynamic_span = {

occurs_dynamic_capacity : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option;

occurs_dynamic_capacity_min : int Cobol_common.Srcloc.TYPES.with_loc option;

occurs_dynamic_capacity_max : int Cobol_common.Srcloc.TYPES.with_loc option;

occurs_dynamic_initialized : bool Cobol_common.Srcloc.TYPES.with_loc;

}

and condition_names = condition_name Cobol_common.Srcloc.TYPES.with_loc list

Named conditions a.k.a. level 88 items

and condition_name = {

condition_name_qualname : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc;

condition_name_item : Cobol_ptree.condition_name_item;

}

and record_renamings = record_renaming Cobol_common.Srcloc.TYPES.with_loc list

Note: RENAMES could be represented by simply adding an (optional, non-constant) offset to redefinitions (and use group layouts with FILLERs throughout to forbid using the new name as a qualifier).

Such a representation would be much more general than what typical COBOL data definitions allow; in particular, one could have "shifted" redefintions of any non-01 group item.

However, we keep the distinction between RENAMES and REDEFINES to better match said typical COBOL, and possibly allow more detailed error reporting.

and record_renaming = {

renaming_name : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc;

renaming_layout : renamed_item_layout;

renaming_offset : Cobol_data__.Data_memory.size;

renaming_size : Cobol_data__.Data_memory.size;

renaming_from : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc;

renaming_thru : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc option;

renaming_has_definition_issues : bool;

}

and renamed_item_layout =

| Renamed_elementary of {

usage : usage;

}

| Renamed_struct of {

subfields : item_definitions;

}

type data_definition =

| Data_field of {

record : record;

def : field_definition Cobol_common.Srcloc.TYPES.with_loc;

}

| Data_renaming of {

record : record;

def : record_renaming Cobol_common.Srcloc.TYPES.with_loc;

}

| Data_condition of {

record : record;

field : field_definition Cobol_common.Srcloc.TYPES.with_loc;

def : condition_name Cobol_common.Srcloc.TYPES.with_loc;

}

| Table_index of {

record : record;

table : table_definition Cobol_common.Srcloc.TYPES.with_loc;

qualname : Cobol_ptree.qualname Cobol_common.Srcloc.TYPES.with_loc;

}

data_definition provides a direct access to the pair of each item and associated record items. It is used for instance to retrieve data item information from its name.

Diagnostics

type error =

| Invalid of {

loc : Cobol_common.Srcloc.TYPES.srcloc;

stuff : invalid_stuff;

}

| Unsupported of {

loc : Cobol_common.Srcloc.TYPES.srcloc;

stuff : unsupported_stuff;

}

| Overlong_literal of {

loc : Cobol_common.Srcloc.TYPES.srcloc;

literal_string : string;

max_length : int;

}

and invalid_stuff =

| Character_in_literal of {

literal_class : literal_class;

char : char;

}

and literal_class =

| Boolean

| Fixed

| Floating

| Hexadecimal

| Integer

and unsupported_stuff =

| Figurative_constant : 'x Cobol_ptree.figurative -> unsupported_stuff

| National_literal

| Concatenation_of_literals

type errors = error NEL.t