Module Cobol_data.Picture

module TYPES : sig ... end

type t = TYPES.picture

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val compare : t -> t -> Ppx_deriving_runtime.int

val pp_detailed_category : TYPES.category Pretty.printer

val pp_category : TYPES.category Pretty.printer

val pp_category_name : TYPES.category Pretty.printer

val pp_picture_symbols : TYPES.symbols list Pretty.printer

val is_edited : TYPES.picture -> bool

is_edited pic indicates whether the given picture string represents an edited item

val is_boolean : TYPES.picture -> bool

val is_national : TYPES.picture -> bool

val is_numeric : TYPES.picture -> bool

is_numeric pic indicates whether the given picture string is of class numeric

val is_signed_numeric : TYPES.picture -> bool

val data_size : TYPES.picture -> int

Size of the underlying data; corresponds to the number of "characters" for usage DISPLAY

val size : TYPES.picture -> int

display size, after editions; corresponds to "size" in standards

val of_string : TYPES.config -> string -> (TYPES.picture, (TYPES.error * (int * int)) list * TYPES.picture) Stdlib.result

val alphanumeric : size:int -> TYPES.picture

val national : size:int -> TYPES.picture

val boolean : int -> TYPES.picture

val digits : int -> TYPES.picture

val fixed_numeric : ?basics:TYPES.basic_edition list -> ?floating:TYPES.floating_insertion -> ?with_sign:bool -> int -> int -> TYPES.picture

module Make (Config : Cobol_config.T) (Env : TYPES.ENV) : sig ... end

val pp_error : TYPES.error Pretty.printer

val rev_errors_with_loc : loc:Cobol_common.Srcloc.TYPES.srcloc -> (TYPES.error * (int * int)) list -> TYPES.error Cobol_common.Srcloc.TYPES.with_loc list

val error_diagnostics : loc:Cobol_common.Srcloc.TYPES.srcloc -> (TYPES.error * (int * int)) list -> Cobol_common.Diagnostics.diagnostics

val pp_meaning_of_precedence_index : decimal_char:char -> Stdlib.Format.formatter -> int -> unit

val unit_test : ?config:TYPES.config -> expect:string -> string -> bool

Verifies that the picture string is interpreted as `expect`, i.e. the result of `pp_picture`. If not, displays the difference on stderr and returns `false`